refactor(osn/api): decompose auth.ts — Phase 1 (pure helpers) (#236)
